+++ peewee.py
@@ -4170,13 +4170,14 @@ class SqliteDatabase(Database):
     json_methods = SqliteJSONMethods
 
     def __init__(self, database, pragmas=None, regexp_function=False,
-                 rank_functions=False, *args, **kwargs):
+                 rank_functions=False, lock_type=None, *args, **kwargs):
         isolation = kwargs.pop('isolation_level', None)
         if isolation is not None:
             raise ImproperlyConfigured('isolation_level must be None when '
                                        'using peewee.')
 
         self._pragmas = pragmas or ()
+        self._lock_type = lock_type
         super(SqliteDatabase, self).__init__(database, *args, **kwargs)
         self._aggregates = {}
         self._collations = {}
@@ -4404,6 +4405,7 @@ class SqliteDatabase(Database):
         return True
 
     def begin(self, lock_type=None):
+        lock_type = lock_type or self._lock_type
         statement = 'BEGIN %s' % lock_type if lock_type else 'BEGIN'
         self.execute_sql(statement)